nvme: Translate CPU physical addresses to PCIe bus addresses for DMA
Extend the NVMe driver to translate CPU physical addresses to PCIe bus addresses as seen from the perspective of the NVMe. This is necessary for platforms where there is no 1:1 mapping between PCIe bus addresses and CPU physical addresses (e.g., when an MMU or PCIe address window remapping is present).